fix: separate restaurant and accounts functionality
