+++ core/urls.py
@@ -20,5 +20,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path("accounts/", include("accounts.urls", namespace="accounts")),
     path("", include("restaurant.urls")),
 ]
