+++ restaurant/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Cook, Dish, DishType
+from .models import Dish, DishType
 
-admin.site.register(Cook)
 admin.site.register(Dish)
 admin.site.register(DishType)
