+++ restaurant/forms.py
@@ -1,6 +1,5 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm
-from .models import Dish, DishType, Cook
+from .models import Dish, DishType
 
 
 class DishForm(forms.ModelForm):
@@ -25,16 +24,3 @@ class DishTypeForm(forms.ModelForm):
                 attrs={"class": "form-control", "placeholder": "Enter dish type name"}
             )
         }
-
-
-class CookCreationForm(UserCreationForm):
-    years_of_experience = forms.IntegerField(min_value=0)
-
-    class Meta(UserCreationForm.Meta):
-        model = Cook
-        fields = UserCreationForm.Meta.fields + (
-            "first_name",
-            "last_name",
-            "email",
-            "years_of_experience",
-        )
