+++ restaurant/models.py
@@ -1,14 +1,4 @@
 from django.db import models
-from django.contrib.auth.models import AbstractUser
-
-
-class Cook(AbstractUser):
-    years_of_experience = models.PositiveIntegerField(default=0)
-    first_name = models.CharField(max_length=150)
-    last_name = models.CharField(max_length=150)
-
-    def __str__(self):
-        return f"{self.first_name} {self.last_name}"
 
 
 class DishType(models.Model):
@@ -23,7 +13,7 @@ class Dish(models.Model):
     description = models.TextField()
     price = models.DecimalField(max_digits=10, decimal_places=2)
     dish_type = models.ForeignKey(DishType, on_delete=models.CASCADE)
-    cooks = models.ManyToManyField(Cook, related_name="dishes")
+    cooks = models.ManyToManyField("accounts.Cook", related_name="dishes")
 
     def __str__(self):
         return self.name
