+++ restaurant/tests.py
@@ -1,20 +1,13 @@
 from django.test import TestCase, Client
 from django.urls import reverse
 from django.contrib.auth import get_user_model
-from .models import Dish, DishType, Cook
-from .forms import DishForm, DishTypeForm, CookCreationForm
+from .models import Dish, DishType
+from .forms import DishForm, DishTypeForm
 from decimal import Decimal
 
 
 class ModelTests(TestCase):
     def setUp(self):
-        self.cook = get_user_model().objects.create_user(
-            username="testcook",
-            password="testpass123",
-            first_name="Test",
-            last_name="Cook",
-            years_of_experience=5,
-        )
         self.dish_type = DishType.objects.create(name="Main Course")
         self.dish = Dish.objects.create(
             name="Test Dish",
@@ -22,10 +15,6 @@ class ModelTests(TestCase):
             price=Decimal("9.99"),
             dish_type=self.dish_type,
         )
-        self.dish.cooks.add(self.cook)
-
-    def test_cook_str(self):
-        self.assertEqual(str(self.cook), "Test Cook")
 
     def test_dish_type_str(self):
         self.assertEqual(str(self.dish_type), "Main Course")
@@ -36,9 +25,6 @@ class ModelTests(TestCase):
 
 class FormTests(TestCase):
     def setUp(self):
-        self.cook = get_user_model().objects.create_user(
-            username="testcook", password="testpass123"
-        )
         self.dish_type = DishType.objects.create(name="Main Course")
 
     def test_dish_form_valid(self):
@@ -47,7 +33,7 @@ class FormTests(TestCase):
             "description": "Description",
             "price": "10.99",
             "dish_type": self.dish_type.id,
-            "cooks": [self.cook.id],
+            "cooks": [],
         }
         form = DishForm(data=form_data)
         self.assertTrue(form.is_valid())
@@ -57,26 +43,10 @@ class FormTests(TestCase):
         form = DishTypeForm(data=form_data)
         self.assertTrue(form.is_valid())
 
-    def test_cook_creation_form_valid(self):
-        form_data = {
-            "username": "newcook",
-            "password1": "testpass123",
-            "password2": "testpass123",
-            "first_name": "New",
-            "last_name": "Cook",
-            "email": "cook@test.com",
-            "years_of_experience": 3,
-        }
-        form = CookCreationForm(data=form_data)
-        self.assertTrue(form.is_valid())
-
 
 class ViewTests(TestCase):
     def setUp(self):
         self.client = Client()
-        self.cook = get_user_model().objects.create_user(
-            username="testcook", password="testpass123"
-        )
         self.dish_type = DishType.objects.create(name="Main Course")
         self.dish = Dish.objects.create(
             name="Test Dish",
@@ -84,13 +54,15 @@ class ViewTests(TestCase):
             price=Decimal("9.99"),
             dish_type=self.dish_type,
         )
-        self.dish.cooks.add(self.cook)
 
-    def test_login_required(self):
+    def test_dish_list_view_requires_login(self):
         response = self.client.get(reverse("restaurant:dish_list"))
-        self.assertEqual(response.status_code, 302)
+        self.assertEqual(response.status_code, 302)  # Redirect to login
 
-    def test_dish_list_view(self):
+    def test_dish_list_view_logged_in(self):
+        cook = get_user_model().objects.create_user(
+            username="testcook", password="testpass123"
+        )
         self.client.login(username="testcook", password="testpass123")
         response = self.client.get(reverse("restaurant:dish_list"))
         self.assertEqual(response.status_code, 200)
@@ -98,20 +70,10 @@ class ViewTests(TestCase):
         self.assertContains(response, "Test Dish")
 
     def test_dish_create_view(self):
+        cook = get_user_model().objects.create_user(
+            username="testcook", password="testpass123"
+        )
         self.client.login(username="testcook", password="testpass123")
         response = self.client.get(reverse("restaurant:dish_create"))
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, "restaurant/dish_form.html")
-
-    def test_dish_edit_view(self):
-        self.client.login(username="testcook", password="testpass123")
-        response = self.client.get(
-            reverse("restaurant:dish_edit", kwargs={"pk": self.dish.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, "restaurant/dish_form.html")
-
-    def test_register_view(self):
-        response = self.client.get(reverse("restaurant:register"))
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, "registration/register.html")
