+++ templates/base.html
@@ -1,64 +1,76 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Restaurant Kitchen Management</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
-</head>
-<body>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+  </head>
+  <body>
     <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-        <div class="container">
-            <a class="navbar-brand" href="{% url 'restaurant:dish_list' %}">Kitchen Management</a>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav me-auto">
-                    {% if user.is_authenticated %}
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'restaurant:dishes' %}">Dishes</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'restaurant:dish_type_list' %}">Dish Types</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'restaurant:cook_list' %}">Cooks</a>
-                        </li>
-                    {% endif %}
-                </ul>
-                {% if user.is_authenticated %}
-                    <form method="post" action="{% url 'restaurant:logout' %}" class="d-flex">
-                        {% csrf_token %}
-                        <button type="submit" class="btn btn-outline-light">Logout</button>
-                    </form>
-                {% else %}
-                    <ul class="navbar-nav">
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'restaurant:login' %}">Login</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'restaurant:register' %}">Register</a>
-                        </li>
-                    </ul>
-                {% endif %}
-            </div>
+      <div class="container">
+        <a class="navbar-brand" href="{% url 'restaurant:dish_list' %}">Kitchen Management</a>
+        <button
+          class="navbar-toggler"
+          type="button"
+          data-bs-toggle="collapse"
+          data-bs-target="#navbarNav"
+        >
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNav">
+          <ul class="navbar-nav me-auto">
+            {% if user.is_authenticated %}
+            <li class="nav-item">
+              <a class="nav-link" href="{% url 'restaurant:dishes' %}">Dishes</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="{% url 'restaurant:dish_type_list' %}">Dish Types</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="{% url 'accounts:cook_list' %}">Cooks</a>
+            </li>
+            {% endif %}
+          </ul>
+          {% if user.is_authenticated %}
+          <form
+            method="post"
+            action="{% url 'accounts:logout' %}"
+            class="d-flex"
+          >
+            {% csrf_token %}
+            <button type="submit" class="btn btn-outline-light">Logout</button>
+          </form>
+          {% else %}
+          <ul class="navbar-nav">
+            <li class="nav-item">
+              <a class="nav-link" href="{% url 'accounts:login' %}">Login</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="{% url 'accounts:register' %}">Register</a>
+            </li>
+          </ul>
+          {% endif %}
         </div>
+      </div>
     </nav>
 
     <div class="container mt-4">
-        {% if messages %}
-            {% for message in messages %}
-                <div class="alert alert-{{ message.tags }}">
-                    {{ message }}
-                </div>
-            {% endfor %}
-        {% endif %}
-        
-        {% block content %}
-        {% endblock %}
+      {% if messages %}
+          {% for message in messages %}
+              <div class="alert alert-{{ message.tags }}">
+                  {{ message }}
+              </div>
+          {% endfor %}
+      {% endif %}
+      
+      {% block content %}
+      {% endblock %}
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-</body>
+  </body>
 </html>
