+++ templates/registration/login.html
@@ -1,27 +1,28 @@
-{% extends 'base.html' %}
-
-{% block content %}
+{% extends 'base.html' %} {% block content %}
 <div class="row justify-content-center">
-    <div class="col-md-6">
-        <div class="card mt-5">
-            <div class="card-body">
-                <h2 class="text-center mb-4">Login</h2>
-                <form method="post">
-                    {% csrf_token %}
-                    {% for field in form %}
-                    <div class="mb-3">
-                        <label for="{{ field.id_for_label }}" class="form-label">{{ field.label }}</label>
-                        {{ field.errors }}
-                        {{ field }}
-                    </div>
-                    {% endfor %}
-                    <button type="submit" class="btn btn-primary w-100">Login</button>
-                </form>
-                <div class="text-center mt-3">
-                    <p>Don't have an account? <a href="{% url 'restaurant:register' %}">Register here</a></p>
-                </div>
-            </div>
+  <div class="col-md-6">
+    <div class="card mt-5">
+      <div class="card-body">
+        <h2 class="text-center mb-4">Login</h2>
+        <form method="post">
+          {% csrf_token %} {% for field in form %}
+          <div class="mb-3">
+            <label for="{{ field.id_for_label }}" class="form-label"
+              >{{ field.label }}</label
+            >
+            {{ field.errors }} {{ field }}
+          </div>
+          {% endfor %}
+          <button type="submit" class="btn btn-primary w-100">Login</button>
+        </form>
+        <div class="text-center mt-3">
+          <p>
+            Don't have an account?
+            <a href="{% url 'accounts:register' %}">Register here</a>
+          </p>
         </div>
+      </div>
     </div>
+  </div>
 </div>
-{% endblock %} 
+{% endblock %}
