+++ templates/registration/register.html
@@ -18,7 +18,7 @@
         <div class="text-center mt-3">
           <p>
             Already have an account?
-            <a href="{% url 'restaurant:login' %}">Login here</a>
+            <a href="{% url 'accounts:login' %}">Login here</a>
           </p>
         </div>
       </div>
